refactor/naming: Rename 'recovery' file.
Most funcs in there were mutable_data so renamed to that for clarity.
Also added 'client_handler' file for more clarity there too.
