+++ safe_core/src/lib.rs
@@ -64,7 +64,7 @@ mod errors;
 mod network_event;
 
 pub use self::client::{
-    mdata_info, recovery, test_create_balance, AuthActions, Client, ClientKeys, MDataInfo,
+    client_handler, mdata_info, mutable_data, test_create_balance, AuthActions, Client, ClientKeys, MDataInfo,
 };
 #[cfg(feature = "mock-network")]
 pub use self::client::{mock_vault_path, MockConnectionManager as ConnectionManager};
